+++ aiosmtpd/docs/NEWS.rst
@@ -5,12 +5,17 @@
 1.2.3 (aiosmtpd-next)
 =====================
 
+Added
+-----
+* Test for ``SMTP.__init__`` behavior after taking out code that edits TLS Context
+
 Fixed/Improved
 --------------
 * ``handle_exception()`` no longer gets called when the client disconnected (Closes #127, #162)
 * Implement & enforce line-length-limit, thus becoming Compliant with RFC 5321 § 4.5.3.1.6
 * Delay all SMTP Status Code replies during ``DATA`` phase until the phase termination (Closes #9)
 * Now catches ``Controller.factory()`` failure during ``Controller.start()`` (Closes #212)
+* :class:`SMTP` no longer edits user-supplied SSL Context (closes #191)
 
 
 1.2.2 (2020-11-08)